cleaning out stray list
